#61 pushToken이 바뀌면 서버에 업데이트
- 서버에 업데이트
- pref에 업데이트
