+++ app/src/main/java/com/hammer/talkbbokki/presentation/MainActivity.kt
@@ -1,20 +1,29 @@
 package com.hammer.talkbbokki.presentation
 
+import android.annotation.SuppressLint
 import android.os.Bundle
+import android.provider.Settings
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.viewModels
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import com.hammer.talkbbokki.presentation.navigation.TalkbbokkiNavHost
 import com.hammer.talkbbokki.ui.theme.TalkbbokkiTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
+    private val TAG = "MainActivity"
+
+    private val viewModel by viewModels<MainActivityViewModel>()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         loadAd(this)
@@ -32,6 +41,29 @@ class MainActivity : ComponentActivity() {
                 }
             }
         }
+
+        getFirebaseToken()
+    }
+
+    @SuppressLint("HardwareIds")
+    private fun getFirebaseToken() {
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(
+            OnCompleteListener { task ->
+                if (!task.isSuccessful) {
+                    Log.w(TAG, "Fetching FCM registration token failed", task.exception)
+                    return@OnCompleteListener
+                }
+
+                val token = task.result
+                Log.d(TAG, "FCM token : $token")
+                val id = Settings.Secure.getString(
+                    applicationContext.contentResolver,
+                    Settings.Secure.ANDROID_ID
+                )
+                Log.d(TAG, "SSAID : $id")
+                viewModel.saveDeviceToken(id, token)
+            }
+        )
     }
 
     override fun onDestroy() {
